// src/app/core/api/generated/model/rule-entity-response.ts
/**
 * B4RRHH Personnel Administration API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RuleEntityResponse { 
    ruleSystemCode: string;
    ruleEntityTypeCode: string;
    code: string;
    name: string;
    description?: string | null;
    active: boolean;
    startDate: string;
    endDate?: string | null;
}